Создает контроллер board.js
