+++ src/controllers/board.js
@@ -1,0 +1,90 @@
+import {ESC_KEYCODE} from '../constants.js';
+import {RenderPosition, render, replace, remove} from '../utils/render.js';
+import TaskEditComponent from '../components/task-edit.js';
+import TaskComponent from '../components/task.js';
+import SortComponent from '../components/sort.js';
+import TasksContainerComponent from '../components/tasks-container.js';
+import NoTasksComponent from '../components/no-tasks.js';
+import LoadMoreButtonComponent from '../components/load-more-btn.js';
+
+const SHOWING_TASKS_COUNT_ON_START = 8;
+const SHOWING_TASKS_COUNT_BY_BUTTON = 8;
+
+const renderTask = (taskList, task) => {
+  const onEscPress = (evt) => {
+    if (evt.keyCode === ESC_KEYCODE) {
+      replaceEditToTask();
+      document.removeEventListener(`keydown`, onEscPress);
+    }
+  };
+
+  const replaceEditToTask = () => {
+    replace(taskComponent, taskEditComponent);
+  };
+
+  const replaceTaskToEdit = () => {
+    replace(taskEditComponent, taskComponent);
+  };
+
+  const taskComponent = new TaskComponent(task);
+  const taskEditComponent = new TaskEditComponent(task);
+
+  taskComponent.setEditButtonClickHandler(() => {
+    replaceTaskToEdit();
+    document.addEventListener(`keydown`, onEscPress);
+  });
+
+  taskEditComponent.setSubmitHandler((evt) => {
+    evt.preventDefault();
+    replaceEditToTask();
+  });
+
+  render(taskList, taskComponent, RenderPosition.BEFOREEND);
+};
+
+export default class BoardController {
+  constructor(container) {
+    this._container = container;
+
+    this._noTasksComponent = new NoTasksComponent();
+    this._sortComponent = new SortComponent();
+    this._tasksComponent = new TasksContainerComponent();
+    this._loadMoreButtonComponent = new LoadMoreButtonComponent();
+  }
+
+  render(tasks) {
+    const container = this._container.getElement();
+    const isAllTasksArchived = tasks.every((task) => task.isArchive);
+
+    if (isAllTasksArchived) {
+      render(container, new NoTasksComponent(), RenderPosition.BEFOREEND);
+      return;
+    }
+
+    render(container, new SortComponent(), RenderPosition.BEFOREEND);
+    render(container, new TasksContainerComponent(), RenderPosition.BEFOREEND);
+
+    const taskList = container.querySelector(`.board__tasks`);
+
+    let showingTasksCount = SHOWING_TASKS_COUNT_ON_START;
+    tasks.slice(0, showingTasksCount)
+      .forEach((task) => {
+        renderTask(taskList, task);
+      });
+
+    const loadMoreButtonComponent = new LoadMoreButtonComponent();
+    render(container, loadMoreButtonComponent, RenderPosition.BEFOREEND);
+
+    loadMoreButtonComponent.setClickHandler(() => {
+      const prevTasksCount = showingTasksCount;
+      showingTasksCount = showingTasksCount + SHOWING_TASKS_COUNT_BY_BUTTON;
+
+      tasks.slice(prevTasksCount, showingTasksCount)
+        .forEach((task) => renderTask(taskList, task));
+
+      if (showingTasksCount >= tasks.length) {
+        remove(loadMoreButtonComponent);
+      }
+    });
+  }
+}
