+++ src/main.js
@@ -1,51 +1,10 @@
-import {ESC_KEYCODE} from './constants.js';
-import {RenderPosition, render, replace, remove} from './utils/render.js';
 import {tasks} from './mock/task.js';
 import {filters} from './mock/filter.js';
+import {render, RenderPosition} from './utils/render.js';
 import BoardComponent from './components/board.js';
-import SortComponent from './components/sort.js';
-import TasksContainerComponent from './components/tasks-container.js';
-import NoTasksComponent from './components/no-tasks.js';
+import BoardController from './controllers/board.js';
 import FilterComponent from './components/site-filter.js';
-import LoadMoreButtonComponent from './components/load-more-btn.js';
-import TaskEditComponent from './components/task-edit.js';
-import TaskComponent from './components/task.js';
 import SiteMenuComponent from './components/site-menu.js';
-
-const SHOWING_TASKS_COUNT_ON_START = 8;
-const SHOWING_TASKS_COUNT_BY_BUTTON = 8;
-
-const renderTask = (taskList, task) => {
-  const onEscPress = (evt) => {
-    if (evt.keyCode === ESC_KEYCODE) {
-      replaceEditToTask();
-      document.removeEventListener(`keydown`, onEscPress);
-    }
-  };
-
-  const replaceEditToTask = () => {
-    replace(taskComponent, taskEditComponent);
-  };
-
-  const replaceTaskToEdit = () => {
-    replace(taskEditComponent, taskComponent);
-  };
-
-  const taskComponent = new TaskComponent(task);
-  const taskEditComponent = new TaskEditComponent(task);
-
-  taskComponent.setEditButtonClickHandler(() => {
-    replaceTaskToEdit();
-    document.addEventListener(`keydown`, onEscPress);
-  });
-
-  taskEditComponent.setSubmitHandler((evt) => {
-    evt.preventDefault();
-    replaceEditToTask();
-  });
-
-  render(taskList, taskComponent, RenderPosition.BEFOREEND);
-};
 
 const siteMainElement = document.querySelector(`.main`);
 const siteControlElement = siteMainElement.querySelector(`.main__control`);
@@ -56,34 +15,5 @@ render(siteMainElement, new FilterComponent(filters), RenderPosition.BEFOREEND);
 const boardComponent = new BoardComponent();
 render(siteMainElement, boardComponent, RenderPosition.BEFOREEND);
 
-const isAllTasksArchived = tasks.every((task) => task.isArchive);
-
-if (isAllTasksArchived) {
-  render(boardComponent.getElement(), new NoTasksComponent(), RenderPosition.BEFOREEND);
-} else {
-  render(boardComponent.getElement(), new SortComponent(), RenderPosition.BEFOREEND);
-  render(boardComponent.getElement(), new TasksContainerComponent(), RenderPosition.BEFOREEND);
-
-  const taskList = boardComponent.getElement().querySelector(`.board__tasks`);
-
-  let showingTasksCount = SHOWING_TASKS_COUNT_ON_START;
-  tasks.slice(0, showingTasksCount)
-    .forEach((task) => {
-      renderTask(taskList, task);
-    });
-
-  const loadMoreButtonComponent = new LoadMoreButtonComponent();
-  render(boardComponent.getElement(), loadMoreButtonComponent, RenderPosition.BEFOREEND);
-
-  loadMoreButtonComponent.setClickHandler(() => {
-    const prevTasksCount = showingTasksCount;
-    showingTasksCount = showingTasksCount + SHOWING_TASKS_COUNT_BY_BUTTON;
-
-    tasks.slice(prevTasksCount, showingTasksCount)
-      .forEach((task) => renderTask(taskList, task));
-
-    if (showingTasksCount >= tasks.length) {
-      remove(loadMoreButtonComponent);
-    }
-  });
-}
+const boardController = new BoardController(boardComponent);
+boardController.render(tasks);
